performance_tests/performance_requirements.txt

# Performance Testing Requirements
# Install with: pip install -r performance_requirements.txt

# Core testing framework
pytest>=7.0.0
pytest-benchmark>=4.0.0
pytest-html>=3.1.0
pytest-xdist>=3.0.0

# HTTP client and load testing
requests>=2.28.0
locust>=2.14.0

# System monitoring
psutil>=5.9.0

# Image processing for test data
Pillow>=9.0.0
numpy>=1.21.0

# Data analysis and reporting
matplotlib>=3.5.0
pandas>=1.4.0
seaborn>=0.11.0

# Async support
aiohttp>=3.8.0
aiofiles>=0.8.0

# Utilities
tqdm>=4.64.0
click>=8.0.0
pyyaml>=6.0
python-dotenv>=0.19.0

# Optional: For more advanced monitoring
# memory-profiler>=0.60.0
# py-spy>=0.3.0

